backend/requirements.txt
django>=5.1.0
djangorestframework>=3.15.2
django-cors-headers>=4.4.0
langgraph>=0.2.39
langchain-openai>=0.2.14
pydantic>=2.8.2
